add product and remove
