+++ Client/hook/useProducts.tsx
@@ -5,17 +5,22 @@ import { TProduct } from '../models/products'
 
 
 const useProducts = () => {
+
+    
     let { data, error, mutate } = useSWR('/products')
 
     const create = async (product: TProduct) => {
         const products = await add(product);
-        mutate([...data, products])
+        mutate(products)
+        // mutate là nó setstate lại cho mình , mình k phải setState như react nữa
     };
 
     // const getProducts = async (product : TProduct) => {
     //     const products = await getAll();
     //     mutate([...data, products])
     // };
+   // [...data] là lấy tất cả data và cũ
+//    [...data, products] lấy tất cả data cũ và mới 
 
     const remove = async (_id:any) => {
         await removeItem(_id);
