+++ Client/package.json
@@ -13,7 +13,7 @@
     "next": "12.1.6",
     "react": "18.1.0",
     "react-dom": "18.1.0",
-    "react-hook-form": "^7.33.1",
+    "react-hook-form": "^7.34.0",
     "sass": "^1.53.0",
     "swr": "^1.3.0"
   },
