+++ Client/pages/admin/products/index.tsx
@@ -1,31 +1,189 @@
 import Link from 'next/link';
 import React from 'react'
-import AdminLayout from '../../../components/Layout/user';
+import AdminLayout from '../../../components/Layout/admin';
 import useProducts from '../../../hook/useProducts'
 import { TProduct } from '../../../models/products';
 
 
 const ProductsList = () => {
-    const { data, error,  remove  } = useProducts();
+    const { data, error, remove } = useProducts();
     console.log(data)
     if (error) return <div>Failed to load </div>
     if (!data) return <div>Loading....</div>
+    console.log(data)
     return (
         <div>
-            {data.map((item: any, index: any) => (
-            <tr key={index}>
-              <th scope="row">{index + 1}</th>
-              <td>{item.name}</td>
-              <td>{item.price}</td>
-              <td>
-                <button className="btn btn-warning">
-                  <Link href={`/products/edit/${item._id}`}>Edit</Link>
-                </button>
-                <button className="btn btn-danger" onClick={() => remove(item._id)}>Delete</button>
-              </td>
-            </tr>
-          ))}
-            products
+
+
+
+            <div className="p-4 bg-white block sm:flex items-center justify-between border-b border-gray-200 lg:mt-1.5">
+                <div className="mb-1 w-full mt-4">
+                    <div className="mb-6">
+                        <h1 className="text-xl sm:text-2xl font-semibold text-gray-900 ">Product Management</h1>
+                    </div>
+                    <div className="block sm:flex items-center md:divide-x md:divide-gray-100">
+
+                        <div className="flex items-center  w-full">
+                            <Link href={'/admin/products/add'}><button type="button" data-modal-toggle="add-product-modal" className="text-white bg-red-600 hover:bg-yellow-500 focus:ring-4 focus:ring-cyan-200 font-medium inline-flex items-center rounded-lg text-sm px-3 py-2 text-center sm:ml-auto">
+                                <svg className="-ml-1 mr-2 h-6 w-6" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+                                    <path fillRule="evenodd" d="M10 5a1 1 0 011 1v3h3a1 1 0 110 2h-3v3a1 1 0 11-2 0v-3H6a1 1 0 110-2h3V6a1 1 0 011-1z" clipRule="evenodd" />
+                                </svg>
+                                Add product
+                            </button></Link>
+
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div className="flex flex-col">
+                <div className="overflow-x-auto">
+                    <div className="align-middle inline-block min-w-full">
+                        <div className="shadow overflow-hidden">
+                            <table className="table-fixed min-w-full divide-y divide-gray-200">
+                                <thead className="bg-gray-100">
+
+                                    <tr>
+                                        <th scope="col" className="p-4">
+                                            <div className="flex items-center">
+                                                <input id="checkbox-all" aria-describedby="checkbox-1" type="checkbox" className="bg-gray-50 border-gray-300 focus:ring-3 focus:ring-cyan-200 h-4 w-4 rounded" />
+                                                <label htmlFor="checkbox-all" className="sr-only">checkbox</label>
+                                            </div>
+                                        </th>
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            ID
+                                        </th>
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            Product Name
+                                        </th>
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            Image
+                                        </th>
+
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            Price
+                                        </th>
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            Size
+                                        </th>
+                                        <th scope="col" className="p-4 text-left text-xs font-medium text-gray-500 uppercase">
+                                            Description
+                                        </th>
+                                        <th scope="col" className="p-4">
+                                        </th>
+                                    </tr>
+                                </thead>
+                                <tbody className="bg-white divide-y divide-gray-200">
+
+                                    {data.map((item: any, index: any) => (
+                                        <tr className="hover:bg-gray-100" key={index}>
+                                            <td className="p-4 w-4">
+                                                <div className="flex items-center">
+                                                    <input id="checkbox-194556" aria-describedby="checkbox-1" type="checkbox" className="bg-gray-50 border-gray-300 focus:ring-3 focus:ring-cyan-200 h-4 w-4 rounded" />
+                                                    <label htmlFor="checkbox-194556" className="sr-only">checkbox</label>
+                                                </div>
+                                            </td>
+                                            <td className="p-4 whitespace-nowrap text-sm font-normal text-gray-500">
+                                                <div className="text-base font-semibold text-gray-900">{index+1}</div>
+                                            </td>
+                                            <td className="p-4 whitespace-nowrap text-sm font-normal text-gray-500">
+                                                <div className="text-base font-semibold text-gray-900">{item.name}</div>
+                                            </td>
+                                            <td className="p-4 whitespace-nowrap text-base font-medium text-gray-900">
+                                               <img src={item.img} width="150px" alt="" /></td>
+                                            <td className="p-4 whitespace-nowrap text-base font-medium text-gray-900">
+                                                {item.price}</td>
+                                            <td className="p-4 whitespace-nowrap text-base font-medium text-gray-900">
+                                            {item.size}</td>
+                                            <td className="p-4 whitespace-nowrap text-base font-medium text-gray-900">
+                                            {item.description}</td>
+                                            <td className="p-4 whitespace-nowrap space-x-2">
+                                                <Link href={'/id/edit'}><button type="button" data-modal-toggle="product-modal" className="text-white bg-cyan-600 hover:bg-cyan-700 focus:ring-4 focus:ring-cyan-200 font-medium rounded-lg text-sm inline-flex items-center px-3 py-2 text-center">
+                                                    <svg className="mr-2 h-5 w-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+                                                        <path d="M17.414 2.586a2 2 0 00-2.828 0L7 10.172V13h2.828l7.586-7.586a2 2 0 000-2.828z">
+                                                        </path>
+                                                        <path fillRule="evenodd" d="M2 6a2 2 0 012-2h4a1 1 0 010 2H4v10h10v-4a1 1 0 112 0v4a2 2 0 01-2 2H4a2 2 0 01-2-2V6z" clipRule="evenodd" />
+                                                    </svg>
+                                                    Edit item
+                                                </button></Link>
+
+                                                <button onClick={() => remove(item._id)} type="button" data-modal-toggle="delete-product-modal" className="text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm inline-flex items-center px-3 py-2 text-center">
+                                                    <svg className="mr-2 h-5 w-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+                                                        <path fillRule="evenodd" d="M9 2a1 1 0 00-.894.553L7.382 4H4a1 1 0 000 2v10a2 2 0 002 2h8a2 2 0 002-2V6a1 1 0 100-2h-3.382l-.724-1.447A1 1 0 0011 2H9zM7 8a1 1 0 012 0v6a1 1 0 11-2 0V8zm5-1a1 1 0 00-1 1v6a1 1 0 102 0V8a1 1 0 00-1-1z" clipRule="evenodd" />
+                                                    </svg>
+                                                    Delete item
+                                                </button>
+                                            </td>
+                                        </tr>
+                                    ))}
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div className="hidden overflow-x-hidden overflow-y-auto fixed top-4 left-0 right-0 md:inset-0 z-50 justify-center items-center h-modal sm:h-full" id="product-modal" aria-hidden="true">
+                <div className="relative w-full max-w-2xl px-4 h-full md:h-auto">
+                    <div className="bg-white rounded-lg shadow relative">
+                        <div className="flex items-start justify-between p-5 border-b rounded-t">
+                            <h3 className="text-xl font-semibold">
+                                Edit product
+                            </h3>
+                            <button type="button" className="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 ml-auto inline-flex items-center" data-modal-toggle="product-modal">
+                                <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+                                    <path fillRule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clipRule="evenodd" />
+                                </svg>
+                            </button>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div className="overflow-x-hidden overflow-y-auto fixed top-4 left-0 right-0 md:inset-0 z-50 justify-center items-center h-modal sm:h-full hidden" id="add-product-modal" aria-hidden="true">
+                <div className="relative w-full max-w-2xl px-4 h-full md:h-auto">
+                    <div className="bg-white rounded-lg shadow relative">
+                        <div className="flex items-start justify-between p-5 border-b rounded-t">
+                            <h3 className="text-xl font-semibold">
+                                Add product
+                            </h3>
+                            <button type="button" className="text-gray-400 bg-transparent hover:bg-gray-200 hover:text-gray-900 rounded-lg text-sm p-1.5 ml-auto inline-flex items-center" data-modal-toggle="add-product-modal">
+                                <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
+                                    <path fillRule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clipRule="evenodd" />
+                                </svg>
+                            </button>
+                        </div>
+                        <div className="p-6 space-y-6">
+                            <form action="#">
+                                <div className="grid grid-cols-6 gap-6">
+                                    <div className="col-span-6 sm:col-span-3">
+                                        <label htmlFor="product-name" className="text-sm font-medium text-gray-900 block mb-2">Product
+                                            Name</label>
+                                        <input type="text" name="product-name" id="product-name" className="shadow-sm bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-cyan-600 focus:border-cyan-600 block w-full p-2.5" placeholder="Apple Imac 27”" required />
+                                    </div>
+                                    <div className="col-span-6 sm:col-span-3">
+                                        <label htmlFor="category" className="text-sm font-medium text-gray-900 block mb-2">Category</label>
+                                        <input type="text" name="category" id="category" className="shadow-sm bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-cyan-600 focus:border-cyan-600 block w-full p-2.5" placeholder="Electronics" required />
+                                    </div>
+                                    <div className="col-span-6 sm:col-span-3">
+                                        <label htmlFor="brand" className="text-sm font-medium text-gray-900 block mb-2">Brand</label>
+                                        <input type="text" name="brand" id="brand" className="shadow-sm bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-cyan-600 focus:border-cyan-600 block w-full p-2.5" placeholder="Apple" required />
+                                    </div>
+                                    <div className="col-span-6 sm:col-span-3">
+                                        <label htmlFor="price" className="text-sm font-medium text-gray-900 block mb-2">Price</label>
+                                        <input type="number" name="price" id="price" className="shadow-sm bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-cyan-600 focus:border-cyan-600 block w-full p-2.5" placeholder="$2300" required />
+                                    </div>
+                                    <div className="col-span-full">
+                                        <label htmlFor="product-details" className="text-sm font-medium text-gray-900 block mb-2">Product
+                                            Details</label>
+                                        <textarea id="product-details" rows={6} className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-cyan-600 focus:border-cyan-600 block w-full p-4" placeholder="e.g. 3.8GHz 8-core 10th-generation Intel Core i7 processor, Turbo Boost up to 5.0GHz, Ram 16 GB DDR4 2300Mhz" defaultValue={""} />
+                                    </div>
+                                </div>
+                            </form>
+                        </div>
+                        <div className="p-6 border-t border-gray-200 rounded-b">
+                            <button className="text-white bg-cyan-600 hover:bg-cyan-700 focus:ring-4 focus:ring-cyan-200 font-medium rounded-lg text-sm px-5 py-2.5 text-center" type="submit">Add product</button>
+                        </div>
+                    </div>
+                </div>
+            </div>
         </div>
     )
 }
